requirements.txt
fastmcp
kubernetes 